% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_species.R
\name{convert_species}
\alias{convert_species}
\title{generic for convert_species}
\usage{
convert_species(x, homologues, ...)
}
\arguments{
\item{x}{A data.frame or data.frame extension.}

\item{homologues}{A two-column data.frame, the first column containing
the IDs for the source-species, the second containing the IDs from the
target-species.}

\item{...}{Further arguments}
}
\value{
An object of the same type as \code{x}.
}
\description{
Given a dataset and a data-frame that contains ID-mappings from one species
(source) to another (target), return a version of \code{x} where the IDs
have been mapped to the IDs in the target species.
}
