% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{drop_incomplete_cases}
\alias{drop_incomplete_cases}
\title{drop_incomplete_cases}
\usage{
drop_incomplete_cases(.x, .cols = colnames(.x))
}
\arguments{
\item{.x}{A data.frame}

\item{.cols}{A vector subset of the column-names/indices of
\code{.x}.}
}
\description{
Filters an input data.frame, keeping only those rows that have complete data
(ie no NA values) within the specified columns
}
