% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{get_duplicates}
\alias{get_duplicates}
\title{get_duplicates}
\usage{
get_duplicates(x)
}
\arguments{
\item{x}{A vector of some form.}
}
\value{
A vector contain the entries of `x` that are duplicated within
`x`.  The order of `x` is not respected and the returned vector is
typically smaller than `x`.
}
\description{
Determines which entries in an input vector are duplicated within that
vector and returns those __values__.
}
