% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologue_db.R
\name{is_valid_mart}
\alias{is_valid_mart}
\title{is_valid_mart}
\usage{
is_valid_mart(mart, id_type = "entrezgene")
}
\arguments{
\item{mart}{A `mart` object from the `biomaRt` package.}

\item{id_type}{Check that this identifier type is a named
field / column in the `mart` object.}
}
\description{
Checks if a user-provided object is a valid `mart` object from the biomaRt
  package; and if it is, where the user-specified id-type is present as one
  of the attributes (ie, columns) of that mart
}
