% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologue_filters.R
\name{keep_only_one_to_one_homologues}
\alias{keep_only_one_to_one_homologues}
\title{keep_only_one_to_one_homologues}
\usage{
keep_only_one_to_one_homologues(forward_map, reverse_map)
}
\arguments{
\item{forward_map}{A `data.frame` containing colnames `id_sp1` and
`id_sp2`.}

\item{reverse_map}{A `data.frame` containing colnames `id_sp1` and
`id_sp2`.}
}
\description{
Takes two homologue maps, as returned by `map_to_homologues`, and drops any
  species1 to species2 mappings in the `forward_map` that aren't 1:1
  homology mappings. This occurs in two cases: i) if ID1 in species1 maps
  to multiple IDs in species2; or ii) for each ID1 in species1, if any of
  the IDs to which that gene maps (in species2) map back to more than one
  gene in species1.
}
\details{
Starting with `forward_map`, if a non-1:1 mapping is found, no filtering
  takes place, and ID1 is unaffected but ID2 is set to NA.
}
