% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologue_functions.R
\name{map_to_homologues}
\alias{map_to_homologues}
\title{map_to_homologues}
\usage{
map_to_homologues(gene_ids = character(0), dataset_sp1 = NULL,
  dataset_sp2 = NULL, sp1 = "hsapiens", sp2 = "mmusculus",
  idtype_sp1 = "ensembl_gene_id", idtype_sp2 = "ensembl_gene_id",
  host = "www.ensembl.org", mart_name = "ENSEMBL_MART_ENSEMBL",
  one_to_one = FALSE)
}
\arguments{
\item{gene_ids}{A vector of gene identifiers for species `sp1`
and of type `idtype_sp1`.}

\item{dataset_sp1}{The `biomaRt` dataset that houses the data for
species `sp1`.}

\item{dataset_sp2}{The `biomaRt` dataset that houses the data for
species `sp2`.}

\item{sp1}{The name of species 1 (using ensembl prefixes
like `hsapiens` / `mmusculus`).}

\item{sp2}{The name of species 2 (using ensembl prefixes
like `hsapiens` / `mmusculus`).}

\item{idtype_sp1}{The type of identifier used for species `sp1`;
this should match one of the field / column names in the biomaRt dataset
for species `sp1`.}

\item{idtype_sp2}{The type of identifier used for species `sp2`;
this should match one of the field / column names in the biomaRt dataset
for species `sp2`.}

\item{host}{The URL for the website that hosts the biomaRt
dataset.}

\item{mart_name}{The name of the biomaRt database that is to be
used.}

\item{one_to_one}{Boolean. Should the function only return
one-to-one homology mappings?}
}
\value{
A `data.frame` with columns `id_sp1` and `id_sp2`.
}
\description{
Takes gene ids from one species and maps them to gene ids in another
species. Can convert entrez ids and ensembl ids.
}
\details{
Note: This function is used in project `dtg_rnaseq`. But the version there
uses dot-separated argument names. The argnames have been converted to
underscore-separated. If this function-script is copied back into the
`dtg_rnaseq` project you will have to double-check the function-calls.
}
