% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologue_functions.R
\name{map_to_homologues_oneway}
\alias{map_to_homologues_oneway}
\title{map_to_homologues_oneway}
\usage{
map_to_homologues_oneway(gene_ids = character(0), dataset_sp1 = NULL,
  dataset_sp2 = NULL, sp1 = "hsapiens", sp2 = "mmusculus",
  idtype_sp1 = "ensembl_gene_id", idtype_sp2 = "ensembl_gene_id")
}
\arguments{
\item{gene_ids}{A vector of gene identifiers for species `sp1`
and of type `idtype_sp1`.}

\item{dataset_sp1}{The `biomaRt` dataset that houses the data for
species `sp1`.}

\item{dataset_sp2}{The `biomaRt` dataset that houses the data for
species `sp2`.}

\item{sp1}{The name of species 1 (using ensembl prefixes
like `hsapiens` / `mmusculus`).}

\item{sp2}{The name of species 2 (using ensembl prefixes
like `hsapiens` / `mmusculus`).}

\item{idtype_sp1}{The type of identifier used for species `sp1`;
this should match one of the field / column names in the biomaRt dataset
for species `sp1`.}

\item{idtype_sp2}{The type of identifier used for species `sp2`;
this should match one of the field / column names in the biomaRt dataset
for species `sp2`.}
}
\description{
Takes gene ids from one species and maps them to gene ids in another
species. Can convert entrez ids and ensembl ids
}
\details{
--- Non-exported function ---
 --- All inputs should have been validity-checked by map_to_homologues ---

Selects all genes from the biomart dataset for species1 that have
   homologues in species 2
 1) Map input-ids in species 1 to ensembl gene in species 2
 2) Map ensembl gene in species 2 to output-ids in species 2

Map species 1 gene_ids to ensembl-ids in species 2
 - Since dataset_sp1 is the mart for species 1, we don't have to pass the
 sp1, host or mart.name arguments into map_to_ensembl_homologues...
}
