% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologue_functions.R
\name{select_and_filter}
\alias{select_and_filter}
\title{select_and_filter : extracts data from biomaRt, removes missing values}
\usage{
select_and_filter(filters, values, attributes, mart)
}
\arguments{
\item{filters}{Vector of filters for use in a biomaRt select().}

\item{values}{List (if multiple filters used) or vector of
values for use in biomaRt::select().}

\item{attributes}{Vector of field-names for return from a biomaRt
query.}

\item{mart}{A biomaRt `mart` object.}
}
\value{
A data.frame containing columns for each of the `attributes`
  that were selected; and where only complete observations were kept (ie,
  there is no missing values in the returned data).
}
\description{
select_and_filter : extracts data from biomaRt, removes missing values
}
