% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologue_db.R
\name{use_default_mart}
\alias{use_default_mart}
\title{use_default_mart}
\usage{
use_default_mart(sp = "hsapiens", host = "www.ensembl.org",
  mart_name = "ENSEMBL_MART_ENSEMBL")
}
\arguments{
\item{sp}{A species name (eg, "hsapiens", "mmusculus") as
used in biomaRt.}

\item{host}{The name of the site that hosts the mart
database.}

\item{mart_name}{The name of the biomart dataset that should be
used.}
}
\description{
Sets up a default biomaRt `mart` object for use in searches
}
