% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologue_filters.R
\name{which_mappings_are_one_to_many}
\alias{which_mappings_are_one_to_many}
\title{which_mappings_are_one_to_many}
\usage{
which_mappings_are_one_to_many(x, seed_col, target_col, na_rm = TRUE)
}
\arguments{
\item{x}{Some data.frame.}

\item{seed_col}{A single string, one of the column names in x.}

\item{target_col}{A single string, one of the column names in x.}

\item{na_rm}{Boolean. Should NA-containing rows be removed
from the (seed_col, target_col) sub-dataframe of x before calling one-many
status?}
}
\description{
Takes a dataframe, and identifies any entries in a given column
  (`seed_col`) that have more than one associated entry in another column of
  that data-frame (`target_col`).
}
\details{
NA-containing (seed_col, target_col) pairs are removed prior to determining
  the number of mappings if `na_rm=TRUE`.
}
