% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologue_functions.R
\name{map_to_ensembl_homologues_with_biomart}
\alias{map_to_ensembl_homologues_with_biomart}
\title{map_to_ensembl_homologues_with_biomart}
\usage{
map_to_ensembl_homologues_with_biomart(gene_ids = character(0),
  dataset_sp1 = NULL, sp1 = "hsapiens", sp2 = "mmusculus",
  idtype_sp1 = "ensembl_gene_id", host = "www.ensembl.org",
  mart_name = "ENSEMBL_MART_ENSEMBL")
}
\arguments{
\item{gene_ids}{A vector of gene_ids, of format matching
`idtype_sp1`.}

\item{dataset_sp1}{A `biomaRt::Mart` object for the first species.}

\item{sp1}{Ensembl contraction of the first species' name
(eg, hsapiens).}

\item{sp2}{Ensembl contraction of the second species' name
(eg, mmusculus).}

\item{idtype_sp1}{The type of id provided for the input species
(sp1). This should be either "entrezgene" or "ensembl_gene_id".}

\item{host}{The URL for the host-site for the biomart
database.}

\item{mart_name}{The name of the biomart database to be used.}
}
\value{
A data.frame with columns ID.sp1 and
  ENSEMBLGENE.sp2.
}
\description{
A function to map from a set of `gene_ids` in one species (sp1) to the
  ensembl-gene id of their homologues in a separate species (sp2). The
  mappings are performed using biomart databases.
}
