% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologue_functions.R
\name{map_to_ensembl_homologues_with_biomart}
\alias{map_to_ensembl_homologues_with_biomart}
\title{map_to_ensembl_homologues_with_biomart}
\usage{
map_to_ensembl_homologues_with_biomart(gene_ids = character(0),
  dataset_sp1 = NULL, sp2 = "mmusculus", idtype_sp1 = "ensembl_gene_id")
}
\arguments{
\item{gene_ids}{A vector of gene identifiers for species `sp1`
and of type `idtype_sp1`.}

\item{dataset_sp1}{The `biomaRt` dataset that houses the data for
species `sp1`.}

\item{sp2}{The name of species 2 (using ensembl prefixes
like `hsapiens` / `mmusculus`).}

\item{idtype_sp1}{The type of identifier used for species `sp1`;
this should match one of the field / column names in the biomaRt dataset
for species `sp1`.}
}
\value{
A data.frame with columns id_sp1 and
  ensembl_gene_sp2.
}
\description{
A function to map from a set of `gene_ids` in one species (sp1) to the
  ensembl-gene id of their homologues in a separate species (sp2). The
  mappings are performed using biomart databases.
}
